Update agent.py
I hope there's not too many mistakes...
I don't know how are we're providing the information about which providers are sharing the news and which not, because  we need to know that for calculating  the truth likelihood. Is the way I'm using the news is correct (how fitness and time being around are accessible, or how these are going to be?). One more thing that I'm not sure about is if it's correct that the truth likelihood has to be changed in the agent (when calling the function that calculates it) or if just the value has to be returned.